Convert Overall Page Metrics widget files to JS modules.
